Use Int64 for LIMIT/OFFSET to match PostgreSQL's int8 wire type
PostgreSQL's LIMIT/OFFSET placeholders are inferred as int8, and bigint
columns are 64-bit on the wire. Surfacing them as arbitrary-precision
Integer in typedSql and as Int in the query builder meant users had to
sprinkle fromIntegral between the two APIs.

- typedSql: map int8 -> Int64 (was Integer). Matches the wire format
  precisely and covers bigint columns, COUNT(*), row_number(), etc.
- QueryBuilder: change limit/offset (and limitClause/offsetClause) to
  Int64, and encode as int8 in HasqlCompiler.
- Pagination: fromIntegral at the one internal call site so the
  user-facing Pagination record keeps its Int fields.

Polymorphic numeric literals (|> limit 10) still work. Callers binding
pageSize :: Int need fromIntegral once, or switch the binding to Int64.

+++ ihp/IHP/QueryBuilder/Types.hs
@@ -148,8 +148,8 @@ data SQLQuery = SQLQuery
     , distinctOnClause :: !(Maybe Text)
     , whereCondition :: !(Maybe Condition)
     , orderByClause :: ![OrderByClause]
-    , limitClause :: !(Maybe Int)
-    , offsetClause :: !(Maybe Int)
+    , limitClause :: !(Maybe Int64)
+    , offsetClause :: !(Maybe Int64)
     , columns :: ![Text]
     , columnsSql :: !Text
     -- ^ Pre-computed qualified column selector, e.g. @"users.id, users.name"@.
@@ -163,8 +163,8 @@ instance SetField "distinctClause" SQLQuery Bool where setField value sqlQuery =
 instance SetField "distinctOnClause" SQLQuery (Maybe Text) where setField value sqlQuery = sqlQuery { distinctOnClause = value }
 instance SetField "whereCondition" SQLQuery (Maybe Condition) where setField value sqlQuery = sqlQuery { whereCondition = value }
 instance SetField "orderByClause" SQLQuery [OrderByClause] where setField value sqlQuery = sqlQuery { orderByClause = value }
-instance SetField "limitClause" SQLQuery (Maybe Int) where setField value sqlQuery = sqlQuery { limitClause = value }
-instance SetField "offsetClause" SQLQuery (Maybe Int) where setField value sqlQuery = sqlQuery { offsetClause = value }
+instance SetField "limitClause" SQLQuery (Maybe Int64) where setField value sqlQuery = sqlQuery { limitClause = value }
+instance SetField "offsetClause" SQLQuery (Maybe Int64) where setField value sqlQuery = sqlQuery { offsetClause = value }
 
 -- | Type class for default scoping of queries
 class DefaultScope table where
